Available Datasets

These are the datasets available from https://www.nationwide.co.uk. For more details you can see use ntwd_browse to browse Nationwide's House Price Index webpage.

Details

The formatted text before the semicolon indicates the `id `that is used to get the data through ntwd_get. After the semicolon there is a short description provided by nationwide.

UK Series

monthly: UK Monthly Indices (Post '91)

quarterly: UK Quarterly Indices (Post '91)

since_1952: UK House Prices Since 1952

inflation_adjusted: UK House Prices Adjusted for Inflation

UK & regional quarterly series – all properties (data available from 1973 onwards)

regional: Regional Quarterly Indices (Post '73)

seasonal_regional: Seasonally Adjusted Regional Quarterly Indices

UK & regional quarterly series - by property age group (data available from 1973 onwards)

new_prop: New Properties (Post '73)

mod_prop: Modern Properties (Post '73) - discontinued

old_prop: Older Properties (Post '73) - discontinued

not_new_prop: Not New Properties (Post '95)

UK & regional quarterly series - by buyer type (data available from 1983 onwards)

first: First Time Buyer (Post '83)

fowner: Former Owner Occupier (Post '83)

UK & regional quarterly series - by property type (data available from 1991 onwards; regional data ends Q2 2015, UK only thereafter)

terraced: Terraced (Post '91)

flats: Flats (Post '91)

semi_det: Semi Detached (Post '91)

detached: Detached (Post '91)

Affordability Estimates

aftb_ind: First Time Buyer Affordability Indices

aftb_hper: First Time Buyer House Price Earnings Ratios