Download House Price Data from Nationwide

This function scrapes Nationwide's House Price Index webpage and downloads data in tidy format. Available datasets are given an `id`. All available `id` and short descriptions can be viewed via `?datasets`. For more information the user can browse the source webpage through ntwd_browse.

ntwd_get(id)

Arguments

id
which dataset `id`to fetch.

Examples

# \donttest{
ntwd_get("monthly")

#> # A tibble: 2,070 x 3
#>    Date       key                  value
#>    <date>     <chr>                <dbl>
#>  1 1991-01-01 Average House Price 53052.
#>  2 1991-02-01 Average House Price 53497.
#>  3 1991-03-01 Average House Price 52893.
#>  4 1991-04-01 Average House Price 53677.
#>  5 1991-05-01 Average House Price 54386.
#>  6 1991-06-01 Average House Price 55107.
#>  7 1991-07-01 Average House Price 54541.
#>  8 1991-08-01 Average House Price 54041.
#>  9 1991-09-01 Average House Price 53259.
#> 10 1991-10-01 Average House Price 53467.
#> # ... with 2,060 more rows

ntwd_get("terraced")

#> # A tibble: 3,220 x 4
#>    Date       region type   value
#>    <date>     <chr>  <chr>  <dbl>
#>  1 1991-01-01 North  Price 33943.
#>  2 1991-04-01 North  Price 35631.
#>  3 1991-07-01 North  Price 35159.
#>  4 1991-10-01 North  Price 35077.
#>  5 1992-01-01 North  Price 33494.
#>  6 1992-04-01 North  Price 34015.
#>  7 1992-07-01 North  Price 34131.
#>  8 1992-10-01 North  Price 33585.
#>  9 1993-01-01 North  Price 35899.
#> 10 1993-04-01 North  Price 37458.
#> # ... with 3,210 more rows

# For a list of datasets
?ntwd_datasets
# }